[FIX] #58 - 카테고리 색상 인덱스를 저장하는 것으로 변경
- colorList 에서 해당하는 Long 값을 찾아서 index를 가져오려 했으나 Model 클래스에서 Long 을 compose의 Color 타입을 저장하고 있어 검색이 어려움..

+++ app/src/main/java/com/woowahantechcamp/account_book/ui/screen/setting/SettingViewModel.kt
@@ -32,7 +32,7 @@ class SettingViewModel @Inject constructor(
         getAllCategoryItem()
     }
 
-    fun saveCategoryItem(id: Int = -1, type: SettingType, title: String, color: Long) {
+    fun saveCategoryItem(id: Int = -1, type: SettingType, title: String, color: Int) {
         viewModelScope.launch {
             val result = if (id == -1) categoryRepository.addCategory(type, title, color)
             else categoryRepository.updateCategory(id, title, color)
